# openapi/openapi.yaml
components:
  schemas:
    AnswerRequest:
      additionalProperties: false
      description: '``POST /v1/answer`` and ``POST /v1/answer/stream``.'
      properties:
        filters:
          $ref: '#/components/schemas/SearchFilters'
        question:
          maxLength: 4000
          minLength: 1
          title: Question
          type: string
        top_k:
          anyOf:
          - maximum: 20.0
            minimum: 1.0
            type: integer
          - type: 'null'
          title: Top K
      required:
      - question
      title: AnswerRequest
      type: object
    AnswerResponse:
      description: 'The answer of ``POST /v1/answer``.


        ``found`` is false when there is no answer. ``reason`` tells why: ``not_found``
        (the model

        found nothing), ``low_relevance`` (the gate stopped weak retrieval before
        the model),

        ``no_context``, ``unverified`` (the answer had no valid citation), ``blocked``
        or ``refused``

        (guardrails).'
      properties:
        answer:
          title: Answer
          type: string
        citations:
          items:
            $ref: '#/components/schemas/CitationItem'
          title: Citations
          type: array
        found:
          title: Found
          type: boolean
        mode_used:
          title: Mode Used
          type: string
        model:
          title: Model
          type: string
        prompt_version:
          title: Prompt Version
          type: string
        reason:
          title: Reason
          type: string
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
        usage:
          $ref: '#/components/schemas/UsageItem'
        warnings:
          items:
            type: string
          title: Warnings
          type: array
      required:
      - request_id
      - answer
      - found
      - reason
      - citations
      - warnings
      - mode_used
      - model
      - prompt_version
      - usage
      title: AnswerResponse
      type: object
    CitationItem:
      description: A source of the answer. ``ref`` is the number used in the answer
        text.
      properties:
        doc_id:
          title: Doc Id
          type: string
        pages:
          items:
            type: integer
          title: Pages
          type: array
        ref:
          title: Ref
          type: integer
        snippet:
          title: Snippet
          type: string
      required:
      - ref
      - doc_id
      - pages
      - snippet
      title: CitationItem
      type: object
    ErrorBody:
      description: Details of an error.
      properties:
        code:
          title: Code
          type: string
        message:
          title: Message
          type: string
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
      required:
      - code
      - message
      title: ErrorBody
      type: object
    ErrorResponse:
      description: 'The common error format. ``UPSTREAM_UNAVAILABLE`` and ``TIMEOUT``
        make the caller fall

        back.'
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      required:
      - error
      title: ErrorResponse
      type: object
...
    SearchFilters:
      additionalProperties: false
      description: Optional filters. All of them are added next to the access filter,
        never instead of it.
      properties:
        created_from:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Created From
        created_to:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Created To
        doc_type:
          items:
            type: string
          maxItems: 50
          title: Doc Type
          type: array
        tags:
          items:
            type: string
          maxItems: 50
          title: Tags
          type: array
      title: SearchFilters
      type: object
...
    UsageItem:
      description: Token counts of the model call.
      properties:
        input_tokens:
          title: Input Tokens
          type: integer
        output_tokens:
          title: Output Tokens
          type: integer
      required:
      - input_tokens
      - output_tokens
      title: UsageItem
      type: object
  securitySchemes:
    adminToken:
      description: Admin token.
      scheme: bearer
      type: http
    serviceToken:
      description: Service-to-service token. mTLS is enforced at the ingress.
      scheme: bearer
      type: http
info:
  title: Semantic Search Service
  version: 0.3.0
openapi: 3.1.0
paths:
...
  /v1/answer:
    post:
      description: An answer from the documents the user may read, with citations.
      operationId: answer_v1_answer_post
      parameters:
      - &id002
        description: The end user. Accepted only from a service that is allowed to
          send it.
        in: header
        name: X-User-Id
        required: true
        schema:
          maxLength: 256
          type: string
      - &id003
        description: Comma separated groups of the end user.
        in: header
        name: X-User-Groups
        required: false
        schema:
          type: string
      - *id001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      security:
      - serviceToken: []
      summary: Answer
      tags:
      - answer
  /v1/answer/stream:
    post:
      description: 'The same answer as server-sent events: ``token``, ``citations``,
        ``done`` (or ``error``).


        Everything that can fail before the first token (search, rights, gate) answers
        with a normal

        error status, so the Java app can fall back. If ``done`` says ``found: false``
        after tokens

        were sent (the answer failed the citation check or a guardrail), the client
        discards them.'
      operationId: answer_stream_v1_answer_stream_post
      parameters:
      - *id002
      - *id003
      - *id001
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerRequest'
        required: true
      responses:
        '200':
          content:
            text/event-stream: {}
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      security:
      - serviceToken: []
      summary: Answer Stream
      tags:
      - answer
...